+++ server/apps/core/services/__init__.py
@@ -1,0 +1,1 @@
+from .rcrainfo_service import RcrainfoService
