+++ server/apps/trak/serializers/manifest_ser.py
@@ -4,9 +4,9 @@ from typing import Dict
 from rest_framework import serializers
 
 from apps.trak.models import Manifest
+from apps.trak.models.manifest_models import AdditionalInfo
 from apps.trak.serializers.handler_ser import ManifestHandlerSerializer
 
-from ..models.manifest_models import AdditionalInfo
 from .base_ser import TrakBaseSerializer
 from .handler_ser import TransporterSerializer
 from .waste_line_ser import WasteLineSerializer
