+++ server/apps/trak/tests/services/test_rcrainfo_services.py
@@ -5,9 +5,9 @@ import pytest
 from emanifest import RcrainfoClient
 from responses import matchers
 
+from apps.core.services import RcrainfoService
 from apps.trak.models import QuickerSign
 from apps.trak.serializers import QuickerSignSerializer
-from apps.trak.services import RcrainfoService
 
 
 class TestRcrainfoService:
